% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_utils.R
\name{create_map}
\alias{create_map}
\title{This module provides functions helping with Java calls
Create a \code{Map} from a vector of arguments to the \code{of} static method}
\usage{
create_map(c("name", "Hadrien"))
  create_map(row_to_map_args(c("NAME", "AGE"), c("Hadrien", 34)))
}
\arguments{
\item{args_vector}{A vector of argument}
}
\value{
A Java Map
}
\description{
This module provides functions helping with Java calls
Create a \code{Map} from a vector of arguments to the \code{of} static method
}
