% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/__exploration__.R
\name{df}
\alias{df}
\title{The goal of this file is to store experiments with rJava et Trevas
in order to share insights and troubles.
Should be push the GitHub repo as necessary.}
\usage{
df(x, df1, df2, ncp, log = FALSE)
}
\description{
The goal of this file is to store experiments with rJava et Trevas
in order to share insights and troubles.
Should be push the GitHub repo as necessary.
}
