% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{get_engine}
\alias{get_engine}
\title{This is the main entry point to the API.}
\usage{
get_engine()
}
\value{
a script engine
}
\description{
This is the main entry point to the API.
}
\author{
Romain Tailhurat
Provides a simple way to create a VTL script engine.
}
