% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_utils.R
\docType{data}
\name{JAVA_FQN}
\alias{JAVA_FQN}
\title{This module provides functions helping with Java calls}
\format{
An object of class \code{list} of length 2.
}
\usage{
JAVA_FQN
}
\description{
This module provides functions helping with Java calls
}
\keyword{datasets}
