% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_utils.R
\name{list_of}
\alias{list_of}
\title{Create a \code{List} from a vector of arguments using the \code{of} static method.}
\usage{
list_of(...)
}
\arguments{
\item{...}{Arguments}
}
\value{
A Java List
}
\description{
Create a \code{List} from a vector of arguments using the \code{of} static method.
}
