% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_utils.R
\name{row_to_map_args}
\alias{row_to_map_args}
\title{Provided with a list of keys (names of a row) and a list of values (from the row),
this function return a vector that will be arguments to \code{map_of}.}
\usage{
row_to_map_args(c("NAME", "AGE"), c("Hadrien", 34))
  # "NAME"    "Hadrien" "AGE"     "34"
}
\description{
Provided with a list of keys (names of a row) and a list of values (from the row),
this function return a vector that will be arguments to \code{map_of}.
}
