% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{vtl_eval}
\alias{vtl_eval}
\title{This module provide evaluation functions for VTL expressions}
\usage{
vtl_eval(vtl_expression)
}
\description{
This module provide evaluation functions for VTL expressions
}
\author{
Romain Tailhurat
}
