% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{set_context}
\alias{set_context}
\title{This module provide evaluation functions for VTL expressions}
\usage{
set_context(context, obj)
}
\description{
This module provide evaluation functions for VTL expressions
}
\author{
Romain Tailhurat
}
