% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{vtl_eval}
\alias{vtl_eval}
\title{Evaluate a VTL expression, with bindings if necessary;}
\usage{
vtl_eval(vtl_expression, bindings = NULL)
}
\arguments{
\item{vtl_expression}{a string being a valid VTL expression}

\item{bindings}{the provided bindings}
}
\description{
Evaluate a VTL expression, with bindings if necessary;
}
